Changed way of getting images to be photo and not document

+++ internal/model/image.go
@@ -6,6 +6,6 @@ type Image struct {
 	gorm.Model
 	Filename string
 	MimeType string
-	FileSize int64
+	FileSize int
 	PostID   uint
 }
